Implemented basic Data Structs (#11)

+++ dust-game/src/boundary.rs
@@ -1,8 +1,8 @@
 pub trait Boundary {
-    fn get_x(&self) -> u64;
-    fn get_y(&self) -> u64;
-    fn get_length(&self) -> u64;
-    fn get_width(&self) -> u64;
+    fn get_x(&self) -> u32;
+    fn get_y(&self) -> u32;
+    fn get_length(&self) -> u32;
+    fn get_width(&self) -> u32;
 }
 
 pub trait Wall: Boundary {
